requirements.txt
django

djangorestframework

psycopg2

psycopg2-binary

pandas